requirements.txt
unsloth
torch
torchaudio
huggingface_hub
datasets
pandas
soundfile
trl
peft
accelerate
bitsandbytes
xformers
sentencepiece
protobuf
omegaconf
einx
soxr
einopspyyaml
